+++ docutils/test/test_parsers/test_recommonmark/__init__.py
@@ -1,7 +1,0 @@
-"""Optional tests with 3rd party CommonMark parser"""
-
-import unittest
-try:
-    import docutils.parsers.recommonmark_wrapper  # noqa: F401
-except ImportError as err:
-    raise unittest.SkipTest(err)